+++ src/core/mod.rs
@@ -1,0 +1,7 @@
+pub mod config;
+pub mod error;
+pub mod message;
+
+pub use config::GenerationConfig;
+pub use error::ToolError;
+pub use message::{Message, MessageVecExt};
