+++ src/loaders.rs
@@ -1,3 +1,4 @@
+use crate::core::GenerationConfig;
 use hf_hub::api::tokio::Api as HfApi;
 use std::path::PathBuf;
 use tokenizers::Tokenizer;
@@ -71,16 +72,6 @@ impl TokenizerLoader {
     }
 }
 
-#[derive(Clone)]
-pub struct GenerationConfig {
-    pub temperature: Option<f64>,
-    pub top_p: Option<f64>,
-    pub top_k: Option<u64>,
-    pub min_p: Option<f64>,
-    pub repeat_penalty: Option<f32>,
-    pub repeat_last_n: Option<usize>,
-    pub eos_token_ids: Vec<u64>,
-}
 
 pub struct GenerationConfigLoader {
     pub generation_config_file_loader: HfLoader,
