+++ src/models/generation/sampling.rs
@@ -5,6 +5,7 @@
 //! Functionality for modeling sampling strategies and logits processing in text generation
 //! with support for temperature-based sampling, top-k filtering, nucleus sampling (top-p),
 //! and combinations thereof.
+use super::params::GenerationParams;
 use candle_core::{Context, DType, Error, Result, Tensor};
 use rand::{distr::Distribution, SeedableRng};
 
@@ -322,64 +323,4 @@ pub fn initialize_logits_processor(params: &GenerationParams, seed: u64) -> Logi
     LogitsProcessor::from_sampling(seed, sampling)
 }
 
-pub fn apply_repeat_penalty(
-    logits: &Tensor,
-    penalty: f32,
-    context: &[u32],
-) -> candle_core::Result<Tensor> {
-    let device = logits.device();
-    let mut logits = logits.to_dtype(candle_core::DType::F32)?.to_vec1::<f32>()?;
-    let mut already_seen = std::collections::HashSet::new();
-    for token_id in context {
-        if already_seen.contains(token_id) {
-            continue;
-        }
-        already_seen.insert(token_id);
-        if let Some(logit) = logits.get_mut(*token_id as usize) {
-            if *logit >= 0. {
-                *logit /= penalty
-            } else {
-                *logit *= penalty
-            }
-        }
-    }
-    let logits_len = logits.len();
-    Tensor::from_vec(logits, logits_len, device)
-}
 
-/// Generation parameters for language models.
-#[derive(Debug, Clone)]
-pub struct GenerationParams {
-    pub temperature: f64,
-    pub repeat_penalty: f32,
-    pub repeat_last_n: usize,
-    pub seed: u64,
-    pub max_len: usize,
-    pub top_p: f64,   // 0.0..=1.0 ; 0 or 1 means disabled
-    pub top_k: usize, // 0 means disabled
-    pub min_p: f64,   // 0.0..=1.0 ; 0 means disabled
-}
-
-impl GenerationParams {
-    pub fn new(
-        temperature: f64,
-        repeat_penalty: f32,
-        repeat_last_n: usize,
-        seed: u64,
-        max_len: usize,
-        top_p: f64,
-        top_k: usize,
-        min_p: f64,
-    ) -> Self {
-        Self {
-            temperature,
-            repeat_penalty,
-            repeat_last_n,
-            seed,
-            max_len,
-            top_p,
-            top_k,
-            min_p,
-        }
-    }
-}
