+++ src/pipelines/fill_mask_pipeline/mod.rs
@@ -5,4 +5,4 @@ pub mod fill_mask_pipeline_builder;
 pub use fill_mask_pipeline_builder::FillMaskPipelineBuilder;
 pub use fill_mask_model::FillMaskModel;
 
-pub use crate::models::modernbert::ModernBertSize;
+pub use crate::models::ModernBertSize;
