+++ src/pipelines/mod.rs
@@ -1,7 +1,6 @@
 pub(crate) mod utils;
 
-#[doc(hidden)]
-pub use utils::model_cache::global_cache;
+pub mod base;
 
 pub mod fill_mask_pipeline;
 pub mod sentiment_analysis_pipeline;
