+++ src/pipelines/sentiment_analysis_pipeline/mod.rs
@@ -5,6 +5,6 @@ pub mod sentiment_analysis_pipeline_builder;
 pub use sentiment_analysis_model::SentimentAnalysisModel;
 pub use sentiment_analysis_pipeline_builder::SentimentAnalysisPipelineBuilder;
 
-pub use crate::models::modernbert::ModernBertSize;
+pub use crate::models::ModernBertSize;
 
 pub use anyhow::Result;
