+++ src/pipelines/text_generation/tools/tool_error.rs
