+++ src/pipelines/text_generation_pipeline/parser.rs
