+++ src/pipelines/text_generation_pipeline/streaming/completion_stream.rs
