+++ src/pipelines/text_generation_pipeline/text_generation_pipeline_builder.rs
@@ -1,12 +1,11 @@
-use crate::models::quantized_gemma3::{Gemma3Model, Gemma3Size};
-use crate::models::quantized_qwen3::{Qwen3Model, Qwen3Size};
-use crate::pipelines::utils::model_cache::{global_cache, ModelOptions};
+use crate::models::{Gemma3Model, Gemma3Size, Qwen3Model, Qwen3Size};
+use crate::utils::{global_cache, ModelOptions};
 use crate::pipelines::utils::load_device_with;
 
 use super::text_generation_model::TextGenerationModel;
 use super::text_generation_pipeline::TextGenerationPipeline;
 use super::xml_generation_pipeline::XmlGenerationPipeline;
-use super::xml_parser::XmlParserBuilder;
+use super::parser::XmlParserBuilder;
 use candle_core::{CudaDevice, Device};
 
 pub struct TextGenerationPipelineBuilder<M: TextGenerationModel> {
