+++ src/pipelines/zero_shot_classification_pipeline/mod.rs
@@ -5,6 +5,6 @@ pub mod zero_shot_classification_pipeline_builder;
 pub use zero_shot_classification_model::ZeroShotClassificationModel;
 pub use zero_shot_classification_pipeline_builder::ZeroShotClassificationPipelineBuilder;
 
-pub use crate::models::modernbert::ModernBertSize;
+pub use crate::models::ModernBertSize;
 
 pub use anyhow::Result;
