+++ src/utils/mod.rs
@@ -1,0 +1,3 @@
+pub mod cache;
+
+pub use cache::{global_cache, ModelCache, ModelOptions};
